replace arrow with feather & coerce int64 to int
Because the feather package has been abandoned and the same team is developing the arrow package for feather files, we decided to migrate to this package.   
There is a problem in handing int64 data coming from python to R which needed handling and coercion to built-in int data type in r.   
this commit fixes #294