replace feather with arrow package & coerce int64 to int (#22)
* replace arrow with feather & coerce int64 to int

Because the feather package has been abandoned and the same team is developing the arrow package for feather files, we decided to migrate to this package.   
There is a problem in handing int64 data coming from python to R which needed handling and coercion to built-in int data type in r.   
this commit fixes #294

* Handling out of range integer columns

adding a **try/catch** block to the original column converter to fix the issue when having **columns with integer values** out of the **.Machine$integer.max** number